Fix the failure when installing rmt-server success
The added workaround for bsc#1195759 has a problem that when a case
installing rmt-server package successfully, the test would go 'else'
and die with 'zypper in rmt-server failed'. Our case is using the dev
version of rmt-server, which is higher then released version. The soft
failure bsc#1195759 is not existing on our case, but it is die. This
ticket is going to fix this problem.