Add --gpg-auto-import-keys for register WE with cmd
Based on bsc#1196215 and bsc#1174657, the 3rd party repos with untrusted
signing keys cannot get activated with SUSEConnect. Now the problem is
resolved by a new feature of SUSEConnect that it supports to use --gpg-
auto-import-keys. When registering WE, NVIDIA-driver repo will be added
which has untrusted signing key. So need to add "--gpg-auto-import-keys"
for registering WE command.