Move active python2 module to module scc_deregistration
The current workaround still has a problem. As long as it calls
scc_deregistraion, it would install python2 module, for migration target
veriosn is SLES15SP3 cases. The bsc#189543 happens after migration and
deregister the OS, we need to add the workaround when needing to deregister
the OS after migration.